Cody Bench: rename `evaluate-autocomplete` to `cody-bench`
The agent cli has had an `evaluate-autocomplete` subcommand for a while.
This command is very powerful, it allows us create ad-hoc offline evaluations for any Cody
features, not just autocomplete. This PR is a pure refactoring, renaming
this command to `cody-bench` to make it clearer it can be used for more
things than just evaluating autocomplete.
