+++ agent/src/cli/cody-bench/AutocompleteMatcher.test.ts
