+++ agent/src/cli/cody-bench/AutocompleteMatcher.ts
