+++ agent/src/cli/cody-bench/isWindows.ts
