+++ agent/src/cli/cody-bench/matchesGlobPatterns.test.ts
