+++ agent/src/cli/cody-bench/matchesGlobPatterns.ts
