+++ agent/src/cli/cody-bench/testTypecheck.ts
@@ -5,7 +5,7 @@ import path from 'node:path'
 
 import type { TestParameters } from './TestParameters'
 import { Timer } from './Timer'
-import type { EvaluateAutocompleteOptions } from './evaluate-autocomplete'
+import type { CodyBenchOptions } from './cody-bench'
 
 async function runCommand(command: string | undefined, cwd: string): Promise<boolean> {
     return new Promise<boolean>(resolve => {
@@ -31,13 +31,13 @@ async function runVoidCommand(command: string | undefined, cwd: string): Promise
     }
 }
 
-export async function testCleanup(options: EvaluateAutocompleteOptions): Promise<void> {
+export async function testCleanup(options: CodyBenchOptions): Promise<void> {
     if (options.worktree) {
         await runVoidCommand(`git worktree remove ${options.worktree}`, options.workspace)
     }
 }
 
-export async function testInstall(options: EvaluateAutocompleteOptions): Promise<void> {
+export async function testInstall(options: CodyBenchOptions): Promise<void> {
     if (!options.testTypecheck) {
         return
     }
@@ -46,7 +46,7 @@ export async function testInstall(options: EvaluateAutocompleteOptions): Promise
             `options.worktree=${options.worktree} is already defined, expected it to be undefined`
         )
     }
-    options.worktree = await fspromises.mkdtemp(path.join(os.tmpdir(), 'evaluate-autocomplete-'))
+    options.worktree = await fspromises.mkdtemp(path.join(os.tmpdir(), 'cody-bench-'))
     await runVoidCommand('git diff --exit-code', options.workspace)
 
     // Create a git worktree so that we can run parallel evaluations. Without
